refactor: rename test-module to client-cert module
