+++ modules/client-cert/android/build.gradle
@@ -3,11 +3,11 @@ plugins {
   id 'expo-module-gradle-plugin'
 }
 
-group = 'expo.modules.testmodule'
+group = 'expo.modules.clientcert'
 version = '0.1.0'
 
 android {
-  namespace "expo.modules.testmodule"
+  namespace "expo.modules.clientcert"
   defaultConfig {
     versionCode 1
     versionName "0.1.0"
