+++ modules/client-cert/android/src/main/AndroidManifest.xml
