+++ modules/client-cert/src/ClientCertModule.web.ts
@@ -1,0 +1,5 @@
+import { registerWebModule, NativeModule } from 'expo';
+
+class ClientCertModule extends NativeModule<{}> {}
+
+export default registerWebModule(ClientCertModule, 'ClientCertModule');
