--- modules/test-module/expo-module.config.json
@@ -1,9 +1,0 @@
-{
-  "platforms": ["apple", "android", "web"],
-  "apple": {
-    "modules": ["TestModule"]
-  },
-  "android": {
-    "modules": ["expo.modules.testmodule.TestModule"]
-  }
-}
