--- modules/test-module/src/TestModule.web.ts
@@ -1,5 +1,0 @@
-import { registerWebModule, NativeModule } from 'expo';
-
-class TestModule extends NativeModule<{}> {}
-
-export default registerWebModule(TestModule, 'TestModule');
